[Bronze II] Title: 숫자, Time: 16 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -23,7 +23,7 @@ int main()
     if (a == b) cnt = 0;
 
     cout << cnt << "\n";
-    for (int i = a + 1; i < b; ++i)
+    for (ll i = a + 1; i < b; ++i)
         cout << i << " ";
 
     return 0;
